Configurations

Atomic configurations

We define a configuration to be a set of orbitals with their associated occupation (i.e. the number of electron on that orbital). We can represent a particular configuration with an instance of the Configuration type.

AtomicLevels.Configuration — Type.

struct Configuration{<:AbstractOrbital}

Represents a configuration – a set of orbitals and their associated occupation number. Furthermore, each orbital can be in one of the following states: :open, :closed or :inactive.

source

TODO

Interface of AbstractOrbital?

The @c_str and @rc_str string macros can be used to conveniently construct configurations:

AtomicLevels.@c_str — Macro.

@c_str -> Configuration{Orbital}

Construct a Configuration, representing a non-relativistic configuration, out of a string.

julia> c"1s2 2s"
1s² 2s

julia> c"[Kr] 4d10 5s2 4f2"
[Kr]ᶜ 4d¹⁰ 4f² 5s²

source

AtomicLevels.@rc_str — Macro.

@rc_str -> Configuration{RelativisticOrbital}

Construct a Configuration representing a relativistic configuration out of a string.

julia> rc"[Ne] 3s 3p- 3p"
[Ne]ᶜ 3s 3p⁻ 3p

julia> rc"[Ne] 3s 3p-2 3p4"
[Ne]ᶜ 3s 3p⁻² 3p⁴

source

Interface

Various methods exist to query configurations for information.

AtomicLevels.num_electrons — Method.

num_electrons(conf::Configuration) -> Int

Return the number of electrons in the configuration.

julia> num_electrons(c"1s2")
2

julia> num_electrons(rc"[Kr] 5s2 5p-2 5p2")
42

source