Internals

Internals

Note

The functions, methods and types documented here are not part of the public API.

AtomicLevels.MQ — Constant.

const MQ = Union{Int,Symbol}

Defines the possible types that may represent the main quantum number. It can either be an non-negative integer or a Symbol value (generally used to label continuum electrons).

source

AtomicLevels.get_noble_core_name — Method.

get_noble_core_name(config::Configuration)

Returns the name of the noble gas with the most electrons whose configuration still forms the first part of the closed part of config, or nothing if no such element is found.

julia> AtomicLevels.get_noble_core_name(c"[He] 2s2")
He

julia> AtomicLevels.get_noble_core_name(c"1s2c 2s2c 2p6c 3s2c")
Ne

julia> AtomicLevels.get_noble_core_name(c"1s2") === nothing
true

source

AtomicLevels.kappa_to_j — Method.

kappa_to_j(κ::Integer) :: HalfInteger

Calculate the j quantum number corresponding to the κ quantum number.

Note: κ is always an integer.

source

AtomicLevels.kappa_to_ℓ — Method.

kappa_to_ℓ(κ::Integer) :: Integer

Calculate the ℓ quantum number corresponding to the κ quantum number.

Note: κ and ℓ values are always integers.

source

AtomicLevels.rconfigurations_from_orbital — Method.

rconfigurations_from_orbital(orbital::Orbital, occupancy)

Generate all Configurations with relativistic orbitals corresponding to the non-relativistic version of the orbital with a given occupancy.

Examples

julia> rconfigurations_from_orbital(o"3p", 2)
3-element Array{Configuration{RelativisticOrbital},1}:
 3p⁻²
 3p⁻ 3p
 3p²

source

AtomicLevels.rconfigurations_from_orbital — Method.

rconfigurations_from_orbital(n, ℓ, occupancy)

Generate all Configurations with relativistic orbitals corresponding to the non-relativistic orbital with n and ℓ quantum numbers, with given occupancy.

Examples

julia> rconfigurations_from_orbital(3, 1, 2)
3-element Array{Configuration{RelativisticOrbital},1}:
 3p⁻²
 3p⁻ 3p
 3p²

source

AtomicLevels.ℓj_to_kappa — Method.

ℓj_to_kappa(ℓ::Integer, j::Real) :: Integer

Converts a valid (ℓ, j) pair to the corresponding κ value.

Note: there is a one-to-one correspondence between valid (ℓ,j) pairs and κ values such that for j = ℓ ± 1/2, κ = ∓(j + 1/2).

source

Base.:- — Method.

-(configuration::Configuration, orbital::AbstractOrbital[, n=1])

Remove n electrons in the orbital orbital from the configuration configuration. If the orbital had previously been :closed or :inactive, it will now be :open.

source

Base.delete! — Method.

delete!(c::Configuration, o::AbstractOrbital)

Remove the entire subshell corresponding to orbital o from configuration c.

julia> delete!(c"[Ar] 4s2 3d10 4p2", o"4s")
[Ar]ᶜ 3d¹⁰ 4p²

source

Base.fill! — Method.

fill!(c::Configuration)

Sets all the occupancies in configuration c to maximum, as determined by the degeneracy function.

See also: fill

source

Base.fill — Method.

fill(configuration)

Ensure all orbitals are at their maximum occupancy.

source

Base.isless — Method.

isless(a::Orbital, b::Orbital)

Compares the orbitals a and b to decide which one comes before the other in a configuration.

Examples

julia> o"1s" < o"2s"
true

julia> o"1s" < o"2p"
true

julia> o"ks" < o"2p"
false

source