Orbitals

Atomic orbitals

Orbital types

AtomicLevels.Orbital — Type.

struct Orbital{N <: AtomicLevels.MQ} <: AbstractOrbital

Represents an orbital with just the angular quantum number ℓ.

The type parameter has to be such that it can represent a proper main quantum number (i.e. a subtype of AtomicLevels.MQ).

source

AtomicLevels.SpinOrbital — Type.

struct SpinOrbital{O<:Orbital} <: AbstractOrbital

Spin orbitals are fully characterized orbitals, i.e. the quantum numbers n, ℓ, mℓ and ms are all specified.

source

Methods

Base.isless — Function.

isless(a::Orbital, b::Orbital)

Compares the orbitals a and b to decide which one comes before the other in a configuration.

Examples

julia> o"1s" < o"2s"
true

julia> o"1s" < o"2p"
true

julia> o"ks" < o"2p"
false

source

AtomicLevels.degeneracy — Function.

degeneracy(orbital::Orbital)

Returns the degeneracy of orbital which is 2(2ℓ+1)

Examples

julia> degeneracy(o"1s")
2

julia> degeneracy(o"2p")
6

source

AtomicLevels.parity — Method.

parity(orbital::Orbital)

Returns the parity of orbital, defined as (-1)^ℓ.

Examples

julia> parity(o"1s")
even

julia> parity(o"2p")
odd

source

AtomicLevels.symmetry — Function.

symmetry(orbital::Orbital)

Returns the symmetry for orbital which is simply ℓ.

source

AtomicLevels.isbound — Function.

isbound(::Orbital)

Returns true is the main quantum number is an integer, false otherwise.

julia> isbound(o"1s")
true

julia> isbound(o"ks")
false

source

AtomicLevels.mℓrange — Function.

mℓrange(orbital::Orbital)

Returns the range of valid values of mℓ for orbital.

Examples

julia> mℓrange(o"2p")
-1:1

source

AtomicLevels.spin_orbitals — Function.

spin_orbitals(orbital)

Generate all permissible spin-orbitals for a given orbital, e.g. 2p -> 2p ⊗ mℓ = {-1,0,1} ⊗ s = {α,β}

Examples

julia> spin_orbitals(o"2p")
6-element Array{SpinOrbital{Orbital{Int64}},1}:
 2p₋₁α
 2p₋₁β
 2p₀α
 2p₀β
 2p₁α
 2p₁β

source