Term symbols

Term symbols

AtomicLevels provides types and methods to work and determine term symbols. The "Term symbol" and "Angular momentum coupling" Wikipedia articles give a good basic overview of the terminology.

For term symbols in LS coupling, AtomicLevels provides the Term type.

AtomicLevels.Term — Type.

struct Term

Represent a term symbol ${}^{2S+1}L_{J}$ with specific parity in LS-coupling. As determining valid $J$ values is simple for given $S$ and $L$ ($|L - S| \leq J \leq L+S$), it is not specified.

Constructors

Term(L::Real, S::Real, parity::Union{Parity,Integer})

Constructs a Term object with the given $L$ and $S$ quantum numbers and parity. L and S both have to be convertible to HalfIntegers and parity must be of type Parity or ±1.

source

The Term objects can also be constructed with the @T_str string macro.

AtomicLevels.@T_str — Macro.

@T_str -> Term

Constructs a Term object out of its canonical string representation.

julia> T"1S"
¹S

julia> T"4Po"
⁴Pᵒ

julia> T"2[3/2]o"
²[3/2]ᵒ

source

The terms function can be used to generate all possible term symbols. In the case of relativistic orbitals, the term symbols are simply the valid $J$ values, represented with the HalfInteger type.

AtomicLevels.terms — Function.

terms(orb::Orbital, w::Int=one(Int))

Returns a list of valid LS term symbols

source

terms(o::RelativisticOrbital, w = 1) -> Vector{HalfInt}

Returns a sorted list of valid $J$ values of w equivalent jj-coupled particles on orbital o (i.e. o^w).

When there are degeneracies (i.e. multiple states with the same $J$ and $M$ quantum numbers), the corresponding $J$ value is repeated in the output array.

source

AtomicLevels.count_terms — Function.

count_terms(orb, occ, term)

Count how many times term occurs among the valid terms of orb^occ. For example:

julia> count_terms(o"1s", 2, T"1S")
1

source

The L and S quantum numbers are not, in general, sufficient to uniquely identify a term. The IntermediateTerm type allows one to specify an additional quantum number which would uniquely identify the term.

AtomicLevels.IntermediateTerm — Type.

struct IntermediateTerm

Represents a term together with its seniority quantum number.

source

AtomicLevels.intermediate_terms — Function.

intermediate_terms(orb::Orbital, w::Int=one(Int))

Generates all IntermediateTerm for a given non-relativstic orbital orb and occupation w.

source