Internals

Internals

Note

The functions, methods and types documented here are not part of the public API.

AtomicLevels.MQ — Constant.

const MQ = Union{Int,Symbol}

Defines the possible types that may represent the main quantum number. It can either be an non-negative integer or a Symbol value (generally used to label continuum electrons).

source

AtomicLevels.final_terms — Method.

final_terms(ts::Vector{<:Vector{<:Union{Term,Real}}})

Generate all possible final terms from the vector of vectors of individual subshell terms by coupling from left to right.

Examples

julia> ts = [[T"1S", T"3S"], [T"2P", T"2D"]]
2-element Array{Array{Term,1},1}:
 [¹S, ³S]
 [²P, ²D]

julia> AtomicLevels.final_terms(ts)
4-element Array{Term,1}:
 ²P
 ²D
 ⁴P
 ⁴D

source

AtomicLevels.get_noble_core_name — Method.

get_noble_core_name(config::Configuration)

Returns the name of the noble gas with the most electrons whose configuration still forms the first part of the closed part of config, or nothing if no such element is found.

julia> AtomicLevels.get_noble_core_name(c"[He] 2s2")
"He"

julia> AtomicLevels.get_noble_core_name(c"1s2c 2s2c 2p6c 3s2c")
"Ne"

julia> AtomicLevels.get_noble_core_name(c"1s2") === nothing
true

source

AtomicLevels.kappa_to_j — Method.

kappa_to_j(κ::Integer) :: HalfInteger

Calculate the j quantum number corresponding to the κ quantum number.

Note: κ is always an integer.

source

AtomicLevels.kappa_to_ℓ — Method.

kappa_to_ℓ(κ::Integer) :: Integer

Calculate the ℓ quantum number corresponding to the κ quantum number.

Note: κ and ℓ values are always integers.

source

AtomicLevels.mqtype — Method.

mqtype(::Orbital{MQ}) = MQ

Returns the main quantum number type of an Orbital.

source

AtomicLevels.mqtype — Method.

mqtype(::RelativisticOrbital{MQ}) = MQ

Returns the main quantum number type of a RelativisticOrbital.

source

AtomicLevels.orbital_priority — Method.

orbital_priority(fun, orig_cfg, orbitals)

Generate priorities for the substitution orbitals, i.e. the preferred ordering of the orbitals in configurations excited from orig_cfg. fun can optionally transform the labels of substitution orbitals, in which case they will be ordered just after their parent orbital in the source configuration; otherwise they will be appended to the priority list.

source

AtomicLevels.rconfigurations_from_orbital — Method.

rconfigurations_from_orbital(orbital::Orbital, occupancy)

Generate all Configurations with relativistic orbitals corresponding to the non-relativistic version of the orbital with a given occupancy.

Examples

julia> AtomicLevels.rconfigurations_from_orbital(o"3p", 2)
3-element Array{Configuration{RelativisticOrbital{N}} where N,1}:
 3p-²
 3p- 3p
 3p²

source

AtomicLevels.rconfigurations_from_orbital — Method.

rconfigurations_from_orbital(n, ℓ, occupancy)

Generate all Configurations with relativistic orbitals corresponding to the non-relativistic orbital with n and ℓ quantum numbers, with given occupancy.

Examples

julia> AtomicLevels.rconfigurations_from_orbital(3, 1, 2)
3-element Array{Configuration{RelativisticOrbital{N}} where N,1}:
 3p-²
 3p- 3p
 3p²

source

AtomicLevels.sorted — Method.

sorted(cfg::Configuration)

Returns cfg if it is already sorted or a sorted copy otherwise.

source

AtomicLevels.xu_terms — Method.

xu_terms(ℓ, w, p)

Return all term symbols for the orbital ℓʷ and parity p; the term multiplicity is computed using AtomicLevels.Xu.X.

Examples

julia> AtomicLevels.xu_terms(3, 3, parity(c"3d3"))
17-element Array{Term,1}:
 ²P
 ²D
 ²D
 ²F
 ²F
 ²G
 ²G
 ²H
 ²H
 ²I
 ²K
 ²L
 ⁴S
 ⁴D
 ⁴F
 ⁴G
 ⁴I

source

AtomicLevels.ℓj_to_kappa — Method.

ℓj_to_kappa(ℓ::Integer, j::Real) :: Integer

Converts a valid (ℓ, j) pair to the corresponding κ value.

Note: there is a one-to-one correspondence between valid (ℓ,j) pairs and κ values such that for j = ℓ ± 1/2, κ = ∓(j + 1/2).

source

Base.:+ — Method.

+(::Configuration, ::Configuration)

Add two configurations together. If both configuration have an orbital, the number of electrons gets added together, but in this case the status of the orbitals must match.

julia> c"1s" + c"2s"
1s 2s

julia> c"1s" + c"1s"
1s²

source

Base.:- — Method.

-(configuration::Configuration, orbital::AbstractOrbital[, n=1])

Remove n electrons in the orbital orbital from the configuration configuration. If the orbital had previously been :closed or :inactive, it will now be :open.

source

Base.:== — Method.

==(a::Configuration, b::Configuration)

Tests if configurations a and b are the same, considering orbital occupancy, ordering, and states.

Examples

julia> c"1s 2s" == c"1s 2s"
true

julia> c"1s 2s" == c"1s 2si"
false

julia> c"1s 2s" == c"2s 1s"
false

source

Base.close — Method.

close(c::Configuration)

Return a corresponding configuration where where all the orbitals are marked :closed.

See also: close!

source

Base.count — Method.

count(::Configuration) -> Int

Return the number of electrons in the configuration.

julia> count(c"[Kr] 5s")
37

source

Base.delete! — Method.

delete!(c::Configuration, o::AbstractOrbital)

Remove the entire subshell corresponding to orbital o from configuration c.

julia> delete!(c"[Ar] 4s2 3d10 4p2", o"4s")
[Ar]ᶜ 3d¹⁰ 4p²

source

Base.fill! — Method.

fill!(c::Configuration)

Sets all the occupancies in configuration c to maximum, as determined by the degeneracy function.

See also: fill

source

Base.fill — Method.

fill(c::Configuration)

Returns a corresponding configuration where the orbitals are completely filled (as determined by degeneracy).

See also: fill!

source

Base.filter — Method.

filter(f, c::Configuration) -> Configuration

Filter out the orbitals from configuration c for which the predicate f returns false. The predicate f needs to take three arguments: orbital, occupancy and state.

julia> filter((o,occ,s) -> o.ℓ == 1, c"[Kr]")
2p⁶ᶜ 3p⁶ᶜ 4p⁶ᶜ

source

Base.in — Method.

in(o::AbstractOrbital, c::Configuration) -> Bool

Checks if orbital o is part of configuration c.

julia> in(o"2s", c"1s2 2s2")
true

julia> o"2p" ∈ c"1s2 2s2"
false

source

Base.isless — Method.

isless(a::Orbital, b::Orbital)

Compares the orbitals a and b to decide which one comes before the other in a configuration.

Examples

julia> o"1s" < o"2s"
true

julia> o"1s" < o"2p"
true

julia> o"ks" < o"2p"
false

source

Base.issorted — Method.

issorted(cfg::Configuration)

Tests if the orbitals of cfg is sorted.

source

Base.replace — Method.

replace(conf, a => b[; append=false])

Substitute one electron in orbital a of conf by one electron in orbital b. If conf is unsorted the substitution is performed in-place, unless append, in which case the new orbital is appended instead.

Examples

julia> replace(c"1s2 2s", o"1s" => o"2p")
1s 2p 2s

julia> replace(c"1s2 2s", o"1s" => o"2p", append=true)
1s 2s 2p

julia> replace(c"1s2 2s"s, o"1s" => o"2p")
1s 2s 2p

source

Base.sort — Method.

sort(cfg::Configuration)

Returns a sorted copy of cfg.

source